Implemented autosetting of dialog width based on label text length
The key to the problem is finding the width of the text in pixels,
dialog units, or some other units that can be converted to something
useful. This can be done using the getPreferredSize member function of
the XLayoutConstrains interface provided by the UnoControlFixedText
Service, which provides a Size structure containing the width and height
(in pixels) of a box required to contain the text of the label control.
I figured this out around 2AM this morning during a bout of insomnia. :)

The preferred label width from the XLayoutConstrains interface in a
usable form is then obtained by converting the Size width value to
dialog units (i.e. APPFONT units). After that conversion, the dialog box
width is set to the sum of margins, gaps, icon width, and the preferred
label width (all in APPFONT units).

"Insomnia is a terrible thing to waste."
                    -- Me (Sept 2021)