Improved sizing & positioning of loCustomMessageBox and its components
loDialogBox:
- Updated and added Dialog and Control Size & Position Variables to
improve usability in derived classes.

loCustomMessageBox:
- Eliminated assignment of variables in the constructor since they are
now calculated when the dialog is shown.
- Corrected Name and default Title of dialog in initBox().
- Eliminated default icon definition to centralize icon determination in
a show method and to require explicit icon indication in a call to show;
the icon will be blank if an invalid icon index is provided. This both
prevents displaying an unintended icon and makes clear in the code the
dialog's intended purpose.
- Changed adjustment of the label font from being based on the
ApplicationFontDescriptor to being based on the LabelFontDescriptor.
- Added adjustment of label height based on its text dimensions and
adjustment of the dialog height based on the label height.
- Improved calculation of dialog width and height based on label size.
- Improved button(s) positioning.
- Rearranged code in the main `show` method and added comments in an
attempt to improve code readability.