Eliminated design flaw in loCommonDialogs' access to extension images
ImageProc was originally designed to get the extension identifier from
the extension's description.xml file. This worked while LO Writer was
running from Eclipse so as to install the the extension. It was later
discovered that it didn't work when Writer was later run outside of
Eclipse; dialogs would be displayed without their icons.

To overcome this problem, ImageProc was reworked such that it should be
instantiated at the beginning of an extension using a constructor having
the extension identifier as an argument, whereat the extension
identifier is saved in a static member. This initialization constructor
need not be called in an application running in Eclipse.

Then, ImageProc is instantiated as needed using the constructor with no
arguments and the getURL method to get the URL for a given image
resource.

loCommonDialogs were updated in accordance with the ImageProc changes.